start of register and formatting

+++ src/main/java/io/swagger/model/Entity/UserEntity.java
@@ -17,14 +17,15 @@ public class UserEntity {
     private String name;
     private String email;
     private String password;
-    private Long day_limit;
-    private Long transaction_limit;
+    private int day_limit;
+    private int transaction_limit;
 
-   @ElementCollection(fetch = FetchType.EAGER)
-   private
-   List<UserType> roles;
+    @ElementCollection(fetch = FetchType.EAGER)
+    private
+    List<UserType> roles;
 
-  public UserEntity(){}
+    public UserEntity() {
+    }
 
     public List<UserType> getType() {
         return roles;
@@ -58,19 +59,19 @@ public class UserEntity {
         this.password = password;
     }
 
-    public Long getDay_limit() {
+    public int getDay_limit() {
         return day_limit;
     }
 
-    public void setDay_limit(Long day_limit) {
+    public void setDay_limit(int day_limit) {
         this.day_limit = day_limit;
     }
 
-    public Long getTransaction_limit() {
+    public int getTransaction_limit() {
         return transaction_limit;
     }
 
-    public void setTransaction_limit(Long transaction_limit) {
+    public void setTransaction_limit(int transaction_limit) {
         this.transaction_limit = transaction_limit;
     }
